cold-email- generator/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
google-generativeai==0.3.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
email-validator==2.1.0
composio-openai==0.1.0
